Fix circular import between motools.steps and mozoo.workflows
Move step config classes from mozoo to motools/steps/configs.py to break
circular dependency. Core library code should not depend on application code.

Changes:
- Create motools/steps/configs.py with PrepareDatasetConfig, PrepareTaskConfig, EvaluateModelConfig
- Update motools/steps imports to use local configs
- Uncomment PrepareTaskStep export in motools/steps/__init__.py
- Update mozoo to import and re-export configs for backward compatibility
- Improve validation to always check import paths in __post_init__

Fixes #224
