+++ motools/steps/__init__.py
@@ -3,8 +3,7 @@
 from .base import BaseStep
 from .evaluate_model import EvaluateModelStep
 from .prepare_dataset import PrepareDatasetStep
-
-# from .prepare_task import PrepareTaskStep  # Temporarily commented to avoid circular import
+from .prepare_task import PrepareTaskStep
 from .submit_training import SubmitTrainingStep
 from .train_model import TrainModelStep
 from .wait_for_training import WaitForTrainingStep
@@ -12,7 +11,7 @@ from .wait_for_training import WaitForTrainingStep
 __all__ = [
     "BaseStep",
     "PrepareDatasetStep",
-    # "PrepareTaskStep",  # Temporarily commented
+    "PrepareTaskStep",
     "SubmitTrainingStep",
     "WaitForTrainingStep",
     "TrainModelStep",
