+++ motools/steps/prepare_dataset.py
@@ -2,16 +2,14 @@
 
 import inspect
 from pathlib import Path
-from typing import TYPE_CHECKING, Any, ClassVar
+from typing import ClassVar
 
 from motools.atom import Atom
 from motools.imports import import_function
 from motools.workflow.base import AtomConstructor
 
 from .base import BaseStep
-
-if TYPE_CHECKING:
-    from mozoo.workflows.train_and_evaluate.config import PrepareDatasetConfig  # noqa: F401
+from .configs import PrepareDatasetConfig
 
 
 class PrepareDatasetStep(BaseStep):
@@ -27,11 +25,11 @@ class PrepareDatasetStep(BaseStep):
     name = "prepare_dataset"
     input_atom_types = {}  # No inputs - starts from scratch
     output_atom_types = {"prepared_dataset": "dataset"}
-    config_class: ClassVar[type[Any]] = Any  # Set at runtime to avoid circular import
+    config_class: ClassVar[type[PrepareDatasetConfig]] = PrepareDatasetConfig
 
     async def execute(
         self,
-        config: Any,
+        config: PrepareDatasetConfig,
         input_atoms: dict[str, Atom],
         temp_workspace: Path,
     ) -> list[AtomConstructor]:
